requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2023.11.17
cffi==1.16.0
minio==7.2.3
pycparser==2.21
pycryptodome==3.20.0
python-dotenv==1.0.1
typing_extensions==4.9.0
urllib3==2.1.0
